wip ecr viewer layout
